Copy src directories with dots in them - closes #559
Before, a src folder of, say, public/foo.bar would not copy properly,
because it would be interpreted as a file name, rather than a
directory. This fixes that.

Note that setting the output to a path with a dot in it will still
interpret it as a file. Not sure if there’s any way around that. It
doesn’t know if you mean a file or directory.