Return mix.scripts() to old behavior. - Important
For those of you who are using the Elixir v6 pre-release versions, this
is an important change.

We’re bringing mix.scripts() and mix.scriptsIn() back to how they used
to be. That is, simply concatenate any number of scripts into one.

For Webpack functionality (And ES2015 to ES5 compiling), you will now
use mix.webpack().

So basically, if before you were using mix.browserify(), now you’ll use
mix.webpack(). This will keep the API as consistent as possible with
past versions.